requirements.txt
torch>=2.2.0
torchvision>=0.19
numpy
pillow
tqdm
pyfiglet
pytesseract>=0.3.10

